virtual_teacher_project/requirements.txt
Django==4.2.16
channels==4.0.0
daphne==4.0.0
motor==3.3.2
pymongo==4.6.0
djongo==1.3.6
PyMuPDF==1.23.26
google-generativeai==0.3.2
langchain==0.1.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
bson==0.5.10
channels-redis==4.1.0
